change url

+++ frontend/src/app/Components/product-carousel/product-carousel.component.ts
@@ -7,12 +7,13 @@ import { SearchService } from '../../services/search.service';
 import { IProduct } from '../../models/Iproduct';
 import { Subscription } from 'rxjs';
 import { CategoryService } from '../../services/category.service';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { CartsService } from '../../services/carts.service';
 import { MessageService } from 'primeng/api'; 
 import { ToastModule } from 'primeng/toast';
 import { RippleModule } from 'primeng/ripple';
 import { MatSnackBarModule, MatSnackBar } from '@angular/material/snack-bar';
+import { AuthService } from '../../services/auth/auth.service';
 
 @Component({
   selector: 'app-product-carousel',
@@ -37,8 +38,11 @@ export class ProductCarouselComponent implements OnInit, OnDestroy {
     private categoryService: CategoryService,
     private searchService: SearchService,
     private cartService: CartsService,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    private authService: AuthService, // Inject AuthService here
+    private router: Router // Inject Router here
   ) {}
+  isLogged!:boolean;
 
   ngOnInit(): void {
     this.loadProducts();
@@ -115,6 +119,13 @@ export class ProductCarouselComponent implements OnInit, OnDestroy {
   }
 
   addToCart(product: IProduct): void {
+    // Check if the user is logged in
+    if (!this.authService.isLogged) {
+      // Redirect to login page if not authenticated
+      this.router.navigate(['/login']);
+      return;
+    }
+
     if (this.cartService.isProductInCart(product.id)) {
       this.cartService.removeProductFromCart(product.id);
       this.snackBar.open('Item removed from cart!', 'Close', {
